fixing errors found by Braden in PR 61
