+++ opencsp/app/sofast/lib/SystemSofastFringe.py
@@ -9,11 +9,13 @@ from numpy import ndarray
 import numpy as np
 
 from opencsp.app.sofast.lib.Fringes import Fringes
+from opencsp.app.sofast.lib.ImageCalibrationAbstract import ImageCalibrationAbstract
 from opencsp.app.sofast.lib.MeasurementSofastFringe import MeasurementSofastFringe as Measurement
 from opencsp.common.lib.camera.ImageAcquisitionAbstract import ImageAcquisitionAbstract
 from opencsp.common.lib.deflectometry.ImageProjection import ImageProjection
 from opencsp.common.lib.geometry.Vxyz import Vxyz
 import opencsp.common.lib.tool.exception_tools as et
+import opencsp.common.lib.tool.hdf5_tools as h5
 
 
 class SystemSofastFringe:
@@ -233,6 +235,56 @@ class SystemSofastFringe:
         # Compare to threshold
         if saturation >= thresh:
             warn('Image is {:.2f}% saturated.'.format(saturation * 100), stacklevel=2)
+
+    def run_gray_levels_cal(
+        self,
+        calibration_class: type[ImageCalibrationAbstract],
+        calibration_hdf5_path_name_ext: str = None,
+        on_capturing: Callable = None,
+        on_captured: Callable = None,
+        on_processing: Callable = None,
+        on_processed: Callable[[ImageCalibrationAbstract]] = None,
+    ) -> None:
+        """Runs the projector-camera intensity calibration"""
+
+        # Capture images
+        def _func_0():
+            # Run the "on capturing" callback
+            if on_capturing != None:
+                on_capturing()
+
+            self.run_display_camera_response_calibration(res=10, run_next=self.run_next_in_queue)
+
+            # Run the "on captured" callback
+            if on_captured != None:
+                on_captured()
+
+        # Process data
+        def _func_1():
+            # Run the "on processing" callback
+            if on_processing != None:
+                on_processing()
+
+            # Get calibration images from System
+            calibration_images = self.get_calibration_images()[0]  # only calibrating one camera
+            # Load calibration object
+            calibration = calibration_class.from_data(calibration_images, self.calibration_display_values)
+            # Save calibration object
+            if calibration_hdf5_path_name_ext != None:
+                calibration.save_to_hdf(calibration_hdf5_path_name_ext)
+            # Save calibration raw data
+            data = [self.calibration_display_values, calibration_images]
+            datasets = ['CalibrationRawData/display_values', 'CalibrationRawData/images']
+            if calibration_hdf5_path_name_ext != None:
+                h5.save_hdf5_datasets(data, datasets, calibration_hdf5_path_name_ext)
+            # Run the "on done" callback
+            if on_processed != None:
+                on_processed(calibration)
+            # Continue
+            self.run_next_in_queue()
+
+        self.set_queue([_func_0, _func_1])
+        self.run_next_in_queue()
 
     def capture_mask_images(self, run_next: Callable | None = None) -> None:
         """
